average value of a gcd
given two random numbers from (1,n) average gcd is given with upper bound being the approximate solution for large numbers and lower bound a lower bound for all choices of n which assumes only one use of every prime. The solution is based on the idea that the probability of a number being divisible by a prime sufficiently smaller than it is n is 1/n, and that the average minimum of 2 randomly chosen numbers in n/3. The question is which primes we should use. Since exactly half of a number n's factors are below sqrt(n) we should look at all the primes from 1 to sqrt(n) and since each of these can have an opposite factor we also deal with the same number of primes after sqrt(n). the rest of the primes do not appear enough between (0,n) to assume they have a probability of 1/n so we treat them as if they have a probability of 1/(a/3) of being a factor of the number.